requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
aiosqlite
jinja2
aiofiles
flask
requests

